tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-none-blocked.html

<!doctype html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="style-src 'none';">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <script>
      var t = async_test("Should not stylesheet when style-src is 'none'");
      var t_spv = async_test("Should fire a securitypolicyviolation event");

      document.addEventListener("securitypolicyviolation", t_spv.step_func_done(function(e) {
        assert_equals("style-src", e.violatedDirective);
      }));
    </script>
    <link href="/content-security-policy/style-src/resources/style-src.css" rel=stylesheet type=text/css>
</head>
<body>
    <div id='log'></div>

    <div id="content">Lorem ipsum</div>

    <script>
      t.step(function() {
        var contentEl = document.getElementById("content");
        var marginLeftVal = getComputedStyle(contentEl).getPropertyValue('margin-left');
        assert_false(marginLeftVal == "2px");
        t.done();
      });
    </script>

</body>
</html>
