LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-none-blocked-expected.txt


PASS Should not stylesheet when style-src is 'none'
FAIL Should fire a securitypolicyviolation event assert_equals: expected "style-src" but got "style-src-elem"
Lorem ipsum
